requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
openai>=1.30.0
gitpython>=3.1.43
structlog>=24.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
slowapi>=0.1.9
streamlit>=1.35.0
pyyaml>=6.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
sqlalchemy>=2.0.30
pytest>=8.2.0
pytest-asyncio>=0.23.7
